Remove mount options from RUN directive to support legacy builder
The buildx container build backend is required for the --mount option in a run directive,
but we have buildx disabled in our Codefresh environment to support legacy functionality.
Unfortunately there is no way to enable buildx for a single repository/trigger and I'm not
sure what would break if we turned it on globally.