Fix issue with mismatched data
Because the data we were passing from the build side didn't match the data on the client side exactly (and due to some indexing issues) we were getting a bug where metadata wasn't matching up with post content during hydration. I've standardized how data is passed to components on the client and server side to fix those hydration issues.